chore: switch to rollup
