--- .babelrc
@@ -1,8 +1,0 @@
-{
-  "presets": ["@babel/preset-env"],
-  "plugins": [
-    "@babel/plugin-transform-spread",
-    "transform-react-jsx",
-    "@babel/plugin-proposal-class-properties"
-  ]
-}
