+++ src/components/SliderRail.jsx
