+++ src/components/Tick.jsx
